requirements.txt
attrs==25.4.0
bxml
certifi==2025.10.5
charset-normalizer==3.4.3
click==8.3.0
colorama==0.4.6
contourpy==1.3.3
cycler==0.12.1
dateutil
filelock==3.20.0
fonttools==4.60.1
fsspec==2025.9.0
functorch
idna==3.11
isympy
labelimg==1.8.6
lxml==6.0.2
markupsafe==3.0.3
matplotlib==3.10.7
mpl_toolkits
mpmath==1.3.0
networkx==3.5
numpy==2.2.6
opencv-python==4.12.0.88
packaging==25.0
pillow==11.3.0
pip==25.2
pkg_resources
polars==1.34.0
psutil==7.1.0
pyparsing==3.2.5
PyQt5==5.15.11
pyqt5-sip==12.17.1
python-dateutil==2.9.0.post0
pytoml==0.1.21
pyyaml==6.0.3
requests==2.32.5
scipy==1.16.2
setuptools==65.5.0
six==1.17.0
sympy==1.14.0
thop==2.0.17
torch==2.8.0
torchvision==0.23.0
typing_extensions==4.15.0
ultralytics==8.3.213
urllib3==2.5.0
yaml
